// portfolio_website_setup_vle9j5/convex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
} from "convex/server";
import type * as achievements from "../achievements.js";
import type * as adminAuth from "../adminAuth.js";
import type * as auth from "../auth.js";
import type * as certificates from "../certificates.js";
import type * as education from "../education.js";
import type * as http from "../http.js";
import type * as profile from "../profile.js";
import type * as router from "../router.js";

/**
 * A utility for referencing Convex functions in your app's API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
declare const fullApi: ApiFromModules<{
  achievements: typeof achievements;
  adminAuth: typeof adminAuth;
  auth: typeof auth;
  certificates: typeof certificates;
  education: typeof education;
  http: typeof http;
  profile: typeof profile;
  router: typeof router;
}>;
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;